added more fields in detail

+++ database/models.py
@@ -60,7 +60,7 @@ class Wine(models.Model):
 	producer = models.ForeignKey(Producer, null=True, blank=True, on_delete=models.CASCADE, default=0)
 	name = models.CharField(max_length=200)
 	year = models.IntegerField(blank=True, null=True)
-	alcohol = models.IntegerField(blank=True, null=True)
+	alcohol = models.FloatField(blank=True, null=True)
 	grape = models.ForeignKey(Grape, null=True, blank=True, on_delete=models.CASCADE, default=0)
 	colour = models.ForeignKey(Colour, null=True, blank=True, on_delete=models.CASCADE, default=0)
 	wine_type = models.ForeignKey(WineType, null=True, blank=True, on_delete=models.CASCADE, default=0)
